+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 * v1
   * Initial commit
   * add support for tags
+  * add custom policy support